+++ server/inc/map/generators/MountainMap.hh
@@ -5,7 +5,7 @@
 // Login   <aracthor@epitech.net>
 // 
 // Started on  Fri Oct 24 13:28:15 2014 
-// Last Update Fri Oct 24 13:37:57 2014 
+// Last Update Mon Oct 27 13:36:24 2014 
 //
 
 #ifndef MOUNTAIN_MAP_HH_
@@ -20,7 +20,7 @@ public:
   ~MountainMap();
 
 protected:
-  int	calcHeight(unsigned int x, unsigned y);
+  float	calcHeight(unsigned int x, unsigned int y);
 };
 
 #endif // !MOUNTAIN_MAP_HH_
